Convert chunking_test.rb to Minitest style
